Add files via upload
Finance 1: 
This is a dataset from Dream Housing Finance company, they deal in all home loans. The company wants to automate the loan eligibility process (real-time) based on customer detail provided while filling the online application form. The details are (Gender, Marital Status, Education, Number of Dependents, Income, Loan Amount, Credit History, and others). 
The project will be focused on getting an insight into customers that a potentially eligible for loan, so the company can focus on these customers.

Data set was downloaded Kaggle as a .csv file, on 19/4/2023.